Correct contentIDBase
Content mapping uses the contentIDBase value in the _deconst.json file for the Sphinx project.  For RPC docs, the ContentIDBase values are based on the branch names (master, v11, v12, v13). 

For the engineering handbook, contentIDBase value is the project name: https://github.com/rackerlabs/docs-rpc/eng-handbook/

For configuration, the contentIDbase must be a unique value per Sphinx project build.  By convention, the API doc projects only build one branch, so we used the project name because that guaranteed a unique ID.  For the RPC docs, we used the branch name in the contentIDBase to make the contentIDBase for each branch build unique.